Fixed tasks 3 and 7
in task 3 salary datatype was supposed to be float, not int, and in task 7, companies that didn't make a profit or lost money were not supposed to added up to the overall avg of the conglomerate

+++ Task_3.py
@@ -7,8 +7,8 @@ with open(f"{file}.txt", 'r') as f_obj:
     data = f_obj.readlines()
     for line in data:
         name, salary = line.split()     
-        med_wages+=int(salary)
-        if int(salary) <20000:
+        med_wages+=float(salary)
+        if float(salary) <20000:
             print(name)
     med_wages = med_wages/len(data)
     print(med_wages)
